acc grad

+++ connectomics/io/io_args.py
@@ -173,7 +173,7 @@ def get_args(mode='train', do_output=True):
 
     args.filters = [int(x) for x in args.model_filters.split(',')]
     args.model_pad_mode,args.model_norm_mode,args.model_act_mode = args.model_conv_mode.split(',')
-    args.data_scale = np.array([int(x) for x in args.data_scale.split(',')])
+    args.data_scale = np.array([float(x) for x in args.data_scale.split(',')])
     args.data_invalid_thres = np.array([float(x) for x in args.data_invalid_thres.split(',')])
     args.pre_model_layer = args.pre_model_layer.split('@')
     args.pre_model_layer_select = np.array([int(x) for x in args.pre_model_layer_select.split('@')])
